Move connection code to separate boto3_connection.py
