+++ core/boto3_connection.py
@@ -1,0 +1,24 @@
+#########
+# Copyright (c) 2016 GigaSpaces Technologies Ltd. All rights reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+#  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  * See the License for the specific language governing permissions and
+#  * limitations under the License.
+
+import boto3
+
+
+def connection(aws_conf):
+    return boto3.Session(
+        aws_access_key_id=aws_conf['aws_access_key_id'],
+        aws_secret_access_key=aws_conf['aws_secret_access_key'],
+        region_name=aws_conf['ec2_region_name'],
+        )
